#238 Update sensor-name-cache only when queried
