+++ Hogajama/configuration/local_configuration/01_port_forward_mongo.sh
@@ -1,3 +1,3 @@
-MONGODB=$(oc get pods -o=custom-columns=NAME:.metadata.name -l name=mongodb --no-headers=true)
+MONGODB=$(oc get pods -l name=mongodb --no-headers=true | grep Running | cut -d" " -f1)
 oc port-forward $MONGODB 27017 &
