+++ Hogajama/hogajama-rs/src/main/java/com/gepardec/hogarama/rest/SensorApiImpl.java
@@ -1,11 +1,11 @@
 package com.gepardec.hogarama.rest;
 
 import com.gepardec.hogarama.domain.sensor.SensorDAO;
+import com.gepardec.hogarama.domain.sensor.SensorNamesCache;
 import com.gepardec.hogarama.domain.watering.WateringDAO;
 import com.gepardec.hogarama.domain.watering.WateringData;
 import com.gepardec.hogarama.rest.mapper.SensorMapper;
 import com.gepardec.hogarama.rest.util.DateUtil;
-import com.gepardec.hogarama.service.schedulers.SensorsScheduler;
 
 import javax.enterprise.context.SessionScoped;
 import javax.inject.Inject;
@@ -29,7 +29,7 @@ public class SensorApiImpl implements SensorApi, Serializable {
 	private WateringDAO wateringDAO;
 
 	@Inject
-	private SensorsScheduler sensorsScheduler;
+	private SensorNamesCache sensorsScheduler;
 
 	@Override
 	public Response getAllSensors(SecurityContext securityContext) {
